refactor: move format amount to string utils
