+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/screen/confirm_payment/ConfirmPaymentMapper.kt
@@ -1,22 +1,12 @@
 package net.thechance.mena.wallet.presentation.screen.confirm_payment
 
 import net.thechance.mena.wallet.domain.model.PaymentConfirmation
+import net.thechance.mena.wallet.presentation.utils.formatAmount
 
 fun PaymentConfirmation.toUi(amount: Double) = ConfirmPaymentScreenState.PaymentUiState(
-    amount = formatNumber(amount),
+    amount = formatAmount(amount),
     receiverName = receiverName,
     receiverImage = receiverImg,
     status = status,
-    balance = formatNumber(balance)
+    balance = formatAmount(balance)
 )
-fun formatNumber(number: Double): String {
-    val parts = number.toString().split(".")
-    val integerPart = parts[0].reversed().chunked(3).joinToString(",").reversed()
-    val decimalPart = if (parts.size > 1 && parts[1].toInt() > 1) parts[1] else null
-
-    return if (decimalPart != null && decimalPart.isNotEmpty()) {
-        "$integerPart.${decimalPart.trimEnd('0').ifEmpty { "0" }}"
-    } else {
-        integerPart
-    }
-}
