+++ wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/utils/StringUtils.kt
@@ -16,3 +16,15 @@ fun formatBalance(balance: Double): String {
 
     return "$wholeString.$decimalString"
 }
+
+fun formatAmount(number: Double): String {
+    val parts = number.toString().split(".")
+    val integerPart = parts[0].reversed().chunked(3).joinToString(",").reversed()
+    val decimalPart = if (parts.size > 1 && parts[1].toInt() > 1) parts[1] else null
+
+    return if (decimalPart != null && decimalPart.isNotEmpty()) {
+        "$integerPart.${decimalPart.trimEnd('0').ifEmpty { "0" }}"
+    } else {
+        integerPart
+    }
+}
